+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/forging/ForgeRecipe.java
@@ -1,9 +1,12 @@
 package insane96mcp.iguanatweaksexpanded.module.mining.forging;
 
+import com.google.gson.JsonObject;
 import insane96mcp.iguanatweaksexpanded.setup.client.ITEBookCategory;
 import net.minecraft.core.NonNullList;
 import net.minecraft.core.RegistryAccess;
+import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.GsonHelper;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
@@ -11,6 +14,8 @@ import net.minecraft.world.item.crafting.Recipe;
 import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.RecipeType;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.common.crafting.CraftingHelper;
+import net.minecraftforge.registries.ForgeRegistries;
 
 public class ForgeRecipe implements Recipe<Container> {
     protected final RecipeType<?> type;
@@ -112,5 +117,59 @@ public class ForgeRecipe implements Recipe<Container> {
     @Override
     public ItemStack getToastSymbol() {
         return new ItemStack(Forging.FORGE.item().get());
+    }
+
+    public static class ForgeRecipeSerializer implements RecipeSerializer<ForgeRecipe> {
+        private final CookieBaker<ForgeRecipe> factory;
+
+        public ForgeRecipeSerializer() {
+            this.factory = ForgeRecipe::new;
+        }
+
+        public ForgeRecipe fromJson(ResourceLocation pRecipeId, JsonObject pJson) {
+            ITEBookCategory category = ITEBookCategory.CODEC.byName(GsonHelper.getAsString(pJson, "category", null), ITEBookCategory.FORGE_MISC);
+            Ingredient ingredient = Ingredient.fromJson(pJson.getAsJsonObject("ingredient"));
+            //Forge: Check if primitive string to keep vanilla or an object which can contain a count field.
+            if (!pJson.has("result"))
+                throw new com.google.gson.JsonSyntaxException("Missing result, expected to find a string or object");
+            int ingredientAmount = GsonHelper.getAsInt(pJson, "amount", 1);
+            Ingredient gear = Ingredient.fromJson(pJson.getAsJsonObject("gear"));
+            ItemStack result;
+            if (pJson.get("result").isJsonObject())
+                result = CraftingHelper.getItemStack(GsonHelper.getAsJsonObject(pJson, "result"), true, true);
+            else {
+                String s1 = GsonHelper.getAsString(pJson, "result");
+                ResourceLocation resourcelocation = new ResourceLocation(s1);
+                result = new ItemStack(ForgeRegistries.ITEMS.getValue(resourcelocation));
+            }
+            int smashesRequired = GsonHelper.getAsInt(pJson, "smashes_required");
+            float experience = GsonHelper.getAsFloat(pJson, "experience", 0);
+            return this.factory.create(pRecipeId, category, ingredient, ingredientAmount, gear, result, smashesRequired, experience);
+        }
+
+        public ForgeRecipe fromNetwork(ResourceLocation pRecipeId, FriendlyByteBuf pBuffer) {
+            ITEBookCategory category = pBuffer.readEnum(ITEBookCategory.class);
+            Ingredient ingredient = Ingredient.fromNetwork(pBuffer);
+            int ingredientAmount = pBuffer.readVarInt();
+            Ingredient gear = Ingredient.fromNetwork(pBuffer);
+            ItemStack result = pBuffer.readItem();
+            int smashesRequired = pBuffer.readVarInt();
+            float experience = pBuffer.readFloat();
+            return this.factory.create(pRecipeId, category, ingredient, ingredientAmount, gear, result, smashesRequired, experience);
+        }
+
+        public void toNetwork(FriendlyByteBuf pBuffer, ForgeRecipe pRecipe) {
+            pBuffer.writeEnum(pRecipe.category());
+            pRecipe.getIngredient().toNetwork(pBuffer);
+            pBuffer.writeVarInt(pRecipe.getIngredientAmount());
+            pRecipe.getGear().toNetwork(pBuffer);
+            pBuffer.writeItem(pRecipe.getResultItem(RegistryAccess.EMPTY));
+            pBuffer.writeVarInt(pRecipe.getSmashesRequired());
+            pBuffer.writeFloat(pRecipe.getExperience());
+        }
+
+        public interface CookieBaker<T extends ForgeRecipe> {
+            T create(ResourceLocation pId, ITEBookCategory pCategory, Ingredient ingredient, int ingredientAmount, Ingredient gear, ItemStack pResult, int smashesRequired, float experience);
+        }
     }
 }